malli/pyproject.toml
[tool.poetry]
name = "todo-web"
version = "0.1.0"
description = ""
authors = ["Kalle Ilves <kalle.ilves@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
psycopg2-binary = "^2.9.3"
Flask-SQLAlchemy = "^2.5.1"
python-dotenv = "^0.19.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-dotenv = "^0.5.2"
robotframework = "^4.1.3"
robotframework-seleniumlibrary = "^6.0.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
